feat: Custom hook para corrección y navegación de actividades. #194
