+++ src/student/views/studentPlayActivityView/StudentPlayActivityView.tsx
@@ -1,4 +1,3 @@
-import { useParams, useNavigate } from "react-router-dom";
 import { motion } from "framer-motion";
 import { FaGamepad } from "react-icons/fa";
 import type { ActivityUI } from "../../types/Activity.type";
@@ -8,6 +7,8 @@ import GameRenderer from "../../components/common/StudentGameRenderer/StudentGam
 import LoadingSpinner from "../../../shared/components/LoadingSpinner/LoadingSpinnerComponent";
 import { useActivityStudent } from "../../hooks/useActivityStudentAPI";
 import { useGameManager } from "../../../shared/hooks/games/useGameManager";
+import { useFinishActivity } from "../../hooks/activities/useFinishActivity";
+import { useActivityNavigation } from "../../hooks/activities/useActivityNavigation";
 import styles from "./StudentPlayActivityView.module.css";
 
 interface StudentPlayActivityViewProps {
@@ -17,14 +18,9 @@ interface StudentPlayActivityViewProps {
 const StudentPlayActivityView: React.FC<StudentPlayActivityViewProps> = ({
   activity,
 }) => {
-  const {
-    loading,
-    currentActivity,
-    registerActivityCompleted,
-    refreshStudentDataAfterCompletion,
-  } = useActivityStudent();
-  const { id } = useParams<{ id: string }>();
-  const navigate = useNavigate();
+  const { loading, currentActivity } = useActivityStudent();
+  const { goBackToActivityView } = useActivityNavigation();
+  const { finishActivity } = useFinishActivity();
 
   // EXPO: Registry Pattern: Obtener el hook del juego apropiado automáticamente
   const gameManager = useGameManager(currentActivity?.name || activity?.name);
@@ -40,23 +36,12 @@ const StudentPlayActivityView: React.FC<StudentPlayActivityViewProps> = ({
     },
   };
 
-  const handleGoBack = () => {
-    if (id) {
-      navigate(`/dashboard/student/actividades/${id}/view`);
-    }
-  };
-
   const handleFinishActivity = async () => {
     if (!currentActivity) return;
 
-    await registerActivityCompleted({
-      activityId: currentActivity.id,
-      state: gameManager?.isGameWon ? "APPROVED" : "DISAPPROVED",
-    });
-    await refreshStudentDataAfterCompletion();
+    await finishActivity(!!gameManager?.isGameWon);
 
     gameManager?.resetGame();
-    navigate(`/dashboard/student/actividades/${id}/review`);
   };
 
   if (loading) {
@@ -73,7 +58,7 @@ const StudentPlayActivityView: React.FC<StudentPlayActivityViewProps> = ({
         <h2>Actividad no encontrada</h2>
         <p>No se pudo cargar la información de la actividad.</p>
         <button
-          onClick={() => navigate("/dashboard/student/actividades/list")}
+          onClick={() => goBackToActivityView()}
           className={styles.backButton}
         >
           Volver a actividades
@@ -102,7 +87,7 @@ const StudentPlayActivityView: React.FC<StudentPlayActivityViewProps> = ({
       <StudentActivityFooter
         loading={loading}
         isFormValid={isGameFinished}
-        onBack={handleGoBack}
+        onBack={goBackToActivityView}
         onNext={handleFinishActivity}
         nextButtonText="Finalizar"
         showBackToList={false}
